feat: define frontend command messages
