+++ crates/libtortillas/src/frontend/mod.rs
@@ -4,12 +4,14 @@
 //! snapshots intended for application and UI integrations. Frontends should
 //! prefer these types over actor messages and protocol internals.
 
+mod command;
 mod event;
 mod listener;
 mod publisher;
 mod subscription;
 mod view;
 
+pub use command::{CoreCommand, CoreCommandResult, TorrentCommand};
 pub use event::{CoreEvent, CoreEventKind, FrontendHealth, FrontendHealthLevel};
 pub use listener::{EngineListener, TorrentListener};
 pub use publisher::DEFAULT_EVENT_CAPACITY;
